+++ src/components/ChatContainer/Balloon/ImageContent/ImageContent.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useRef, useState } from 'react';
 import styled from 'styled-components';
 import { Modal } from '@shared';
 import { MessageWithImage } from '@shared/models';
@@ -7,7 +7,7 @@ interface Props {
   image: MessageWithImage['image'];
 }
 
-export const StyledImage = styled.img`
+export const StyledImageContent = styled.img`
   max-width: 100%;
   max-height: 500px;
   cursor: pointer;
@@ -22,12 +22,20 @@ const StyledModalImg = styled.img`
   max-height: 90%;
 `;
 
-export const Image = (props: Props) => {
+export const ImageContent = (props: Props) => {
   const [displayModal, setDisplayModal] = useState(false);
+
+  const styledImageContentRef = useRef<HTMLImageElement>(null);
+
+  const onLoad = () => {
+    const $current = styledImageContentRef.current as HTMLImageElement;
+    $current.style.display = 'block';
+  };
 
   return (
     <>
-      <StyledImage src={props.image} onClick={() => setDisplayModal(true)}/>
+      <StyledImageContent src={props.image} onClick={() => setDisplayModal(true)}
+        onLoad={onLoad} ref={styledImageContentRef} style={{ display: "none" }} />
       <Modal state={[displayModal, setDisplayModal]}>
         <StyledModalImg src={props.image} />
       </Modal>