+++ src/shared/Spinner.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const StyledSpinner = styled.div`
+`;
+
+export const Spinner = () => {
+  
+  return (
+    <StyledSpinner>
+      Loading Image ...
+    </StyledSpinner>
+  );
+};